Refactor: Extract ConverterFactory class
